search-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
strawberry-graphql==0.211.1
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
opensearch-py==2.3.1
